Add zoom level model
