+++ app/core/stores/sequencer/view/index.ts
@@ -3,27 +3,7 @@ import { action, computed, observable } from 'mobx'
 import sequencerLayoutStore from 'core/stores/sequencer/layout'
 import sequencerState from 'core/stores/sequencer/state'
 
-class ZoomLevel {
-  @observable horizontal: number
-  @observable vertical: number
-
-  zoomInterval = 0.1
-
-  constructor(horizontal = 1, vertical = 1) {
-    this.horizontal = horizontal
-    this.vertical = vertical
-  }
-
-  @action
-  zoomInHorizontal() {
-    this.horizontal += this.zoomInterval
-  }
-
-  @action
-  zoomOutHorizontal() {
-    this.horizontal = Math.max(0, (this.horizontal -= this.zoomInterval))
-  }
-}
+import ZoomLevel from 'core/models/zoom-level'
 
 class SequencerViewStore {
   static mobxLoggerConfig = {
